--- projects/storefrontlib/src/common-components/index.ts
@@ -1,2 +1,0 @@
-export * from './star-rating/star-rating.component';
-export * from './star-rating/star-rating.module';
