--- projects/storefrontlib/src/common-components/star-rating/star-rating.component.html
@@ -1,8 +1,0 @@
-<cx-icon
-  *ngFor="let i of [1, 2, 3, 4, 5]"
-  [type]="iconTypes.STAR"
-  class="star"
-  (mouseover)="setRate(i)"
-  (mouseout)="setRate(0)"
-  (click)="saveRate(i)"
-></cx-icon>
