--- projects/storefrontlib/src/common-components/star-rating/star-rating.component.spec.ts
@@ -1,33 +1,0 @@
-import { Component, Input } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { StarRatingComponent } from './star-rating.component';
-
-@Component({
-  selector: 'cx-icon',
-  template: '',
-})
-class MockIconComponent {
-  @Input() type;
-}
-
-describe('StarRatingComponent in product', () => {
-  let starRatingComponent: StarRatingComponent;
-  let fixture: ComponentFixture<StarRatingComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [],
-      declarations: [MockIconComponent, StarRatingComponent],
-      providers: [],
-    }).compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(StarRatingComponent);
-    starRatingComponent = fixture.componentInstance;
-  });
-
-  it('should be created', () => {
-    expect(starRatingComponent).toBeTruthy();
-  });
-});
