--- projects/storefrontlib/src/common-components/star-rating/star-rating.module.ts
@@ -1,11 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { IconModule } from '../../cms-components/misc/index';
-import { StarRatingComponent } from './star-rating.component';
-
-@NgModule({
-  imports: [CommonModule, IconModule],
-  declarations: [StarRatingComponent],
-  exports: [StarRatingComponent],
-})
-export class StarRatingModule {}
