+++ projects/storefrontlib/src/shared/components/star-rating/star-rating.component.html
@@ -1,14 +1,8 @@
-<div class="cx-star-rating" tabindex="0">
-  <ng-template #template let-fill="fill">
-    <span class="star" [class.full]="fill === 100">
-      <span class="half" [style.width.%]="fill">&#9733;</span> &#9733;
-    </span>
-  </ng-template>
-  <ngb-rating
-    [(rate)]="rating"
-    (rateChange)="onTouched(); setRating($event)"
-    [starTemplate]="template"
-    [readonly]="disabled"
-    max="5"
-  ></ngb-rating>
-</div>
+<cx-icon
+  *ngFor="let i of [1, 2, 3, 4, 5]"
+  [type]="iconTypes.STAR"
+  class="star"
+  (mouseover)="setRate(i)"
+  (mouseout)="setRate(0)"
+  (click)="saveRate(i)"
+></cx-icon>
