+++ projects/storefrontlib/src/shared/components/star-rating/star-rating.component.spec.ts
@@ -1,15 +1,24 @@
+import { Component, Input } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NgbRatingConfig, NgbRatingModule } from '@ng-bootstrap/ng-bootstrap';
 import { StarRatingComponent } from './star-rating.component';
+
+@Component({
+  selector: 'cx-icon',
+  template: '',
+})
+class MockIconComponent {
+  @Input() type;
+}
+
 describe('StarRatingComponent in product', () => {
   let starRatingComponent: StarRatingComponent;
   let fixture: ComponentFixture<StarRatingComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [NgbRatingModule],
-      declarations: [StarRatingComponent],
-      providers: [NgbRatingConfig],
+      imports: [],
+      declarations: [MockIconComponent, StarRatingComponent],
+      providers: [],
     }).compileComponents();
   }));
