+++ projects/storefrontlib/src/shared/components/star-rating/star-rating.module.ts
@@ -1,11 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { BootstrapModule } from '../../../lib/bootstrap.module';
-import { StarRatingComponent } from '../star-rating/star-rating.component';
+import { IconModule } from '../../../cms-components/misc/index';
+import { StarRatingComponent } from './star-rating.component';
 
 @NgModule({
-  imports: [CommonModule, FormsModule, ReactiveFormsModule, BootstrapModule],
+  imports: [CommonModule, IconModule],
   declarations: [StarRatingComponent],
   exports: [StarRatingComponent],
 })
